requirements.txt
appdirs==1.4.4
Bottleneck==1.3.5
brotlipy==0.7.0
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.0.4
click==8.0.4
contourpy==1.0.7
cryptography==39.0.1
cycler==0.11.0
Flask==2.2.2
fonttools==4.39.0
gcloud==0.18.3
google-api-core==2.11.0
google-auth==2.16.2
google-cloud-core==2.3.2
google-cloud-storage==2.7.0
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.58.0
httplib2==0.22.0
idna==3.4
itsdangerous==2.0.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.1.1
kiwisolver==1.4.4
MarkupSafe==2.1.1
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
numexpr==2.8.4
numpy
oauth2client==4.1.3
opencv-python==4.7.0.72
packaging==22.0
pandas==1.5.3
Pillow==9.4.0
pip==23.0.1
pooch==1.4.0
protobuf==4.22.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyOpenSSL==23.0.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.7
requests==2.28.1
rsa==4.9
scikit-learn==1.2.2
scipy==1.10.1
setuptools==67.6.1
six==1.16.0
threadpoolctl==2.2.0
urllib3==1.26.14
waitress==2.1.2
Werkzeug==2.2.2
wheel==0.38.4
